# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CounselConnect API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/enrollment-verifications/cor": {
      "post": {
        "tags": [
          "enrollment_verification"
        ],
        "summary": "Student: upload current COR (registration form) PDF",
        "operationId": "upload_cor_api_v1_enrollment_verifications_cor_post",
        "parameters": [
          {
            "name": "counselconnect_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Counselconnect Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_cor_api_v1_enrollment_verifications_cor_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/enrollment-verifications/pending": {
      "get": {
        "tags": [
          "enrollment_verification"
        ],
        "summary": "Counselor: pending queue with applicant details",
        "operationId": "list_pending_api_v1_enrollment_verifications_pending_get",
        "parameters": [
          {
            "name": "counselconnect_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Counselconnect Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PendingItemResponse"
                  },
                  "title": "Response List Pending Api V1 Enrollment Verifications Pending Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/enrollment-verifications/history": {
      "get": {
        "tags": [
          "enrollment_verification"
        ],
        "summary": "Counselor: permanent review record (all applications; optional status filter)",
        "description": "Permanent record of every registration application and its decision.\n\nDecided rows keep decision_at, reviewer, comment, and valid_until.\n`file` is omitted (null) because COR bytes are purged after decisions\nper the privacy boundary.",
        "operationId": "list_history_api_v1_enrollment_verifications_history_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "counselconnect_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Counselconnect Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PendingItemResponse"
                  },
                  "title": "Response List History Api V1 Enrollment Verifications History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/enrollment-verifications/{verification_id}/cor": {
      "get": {
        "tags": [
          "enrollment_verification"
        ],
        "summary": "Counselor: view the uploaded COR PDF",
        "description": "Serves the PDF for the counselor's in-app preview (session cookie).",
        "operationId": "get_cor_pdf_api_v1_enrollment_verifications__verification_id__cor_get",
        "parameters": [
          {
            "name": "verification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Verification Id"
            }
          },
          {
            "name": "counselconnect_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Counselconnect Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/enrollment-verifications/{verification_id}/approve": {
      "post": {
        "tags": [
          "enrollment_verification"
        ],
        "summary": "Counselor: approve the registration application",
        "operationId": "approve_api_v1_enrollment_verifications__verification_id__approve_post",
        "parameters": [
          {
            "name": "verification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Verification Id"
            }
          },
          {
            "name": "counselconnect_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Counselconnect Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ApproveRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/enrollment-verifications/{verification_id}/reject": {
      "post": {
        "tags": [
          "enrollment_verification"
        ],
        "summary": "Counselor: reject the registration application (comment required)",
        "operationId": "reject_api_v1_enrollment_verifications__verification_id__reject_post",
        "parameters": [
          {
            "name": "verification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Verification Id"
            }
          },
          {
            "name": "counselconnect_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Counselconnect Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApproveRequest": {
        "properties": {
          "valid_months": {
            "type": "integer",
            "maximum": 24.0,
            "minimum": 1.0,
            "title": "Valid Months",
            "default": 12
          }
        },
        "type": "object",
        "title": "ApproveRequest"
      },
...
      "Body_upload_cor_api_v1_enrollment_verifications_cor_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_cor_api_v1_enrollment_verifications_cor_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PendingItemResponse": {
        "properties": {
          "verification": {
            "$ref": "#/components/schemas/VerificationResponse"
          },
          "student": {
            "$ref": "#/components/schemas/StudentSummaryResponse"
          },
          "file": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerificationFileMeta"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "verification",
          "student"
        ],
        "title": "PendingItemResponse",
        "description": "One queue/history entry: verification + applicant (+ COR metadata when pending)."
      },
...
      "RejectRequest": {
        "properties": {
          "comment": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "comment"
        ],
        "title": "RejectRequest"
      },
...
      "StudentSummaryResponse": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "middle_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Middle Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "student_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Student Number"
          },
          "campus_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campus Id"
          },
          "program_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Program Id"
          },
          "year_level": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year Level"
          },
          "section": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "email",
          "first_name",
          "middle_name",
          "last_name"
        ],
        "title": "StudentSummaryResponse",
        "description": "Applicant details shown to the reviewer (DFD 1.3 review screen)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerificationFileMeta": {
        "properties": {
          "file_id": {
            "type": "integer",
            "title": "File Id"
          },
          "verification_id": {
            "type": "integer",
            "title": "Verification Id"
          },
          "mime_type": {
            "type": "string",
            "title": "Mime Type"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "cleanup_state": {
            "type": "string",
            "enum": [
              "PENDING",
              "FAILED"
            ],
            "title": "Cleanup State"
          }
        },
        "type": "object",
        "required": [
          "file_id",
          "verification_id",
          "mime_type",
          "size_bytes",
          "expires_at",
          "cleanup_state"
        ],
        "title": "VerificationFileMeta"
      },
      "VerificationResponse": {
        "properties": {
          "verification_id": {
            "type": "integer",
            "title": "Verification Id"
          },
          "student_user_id": {
            "type": "integer",
            "title": "Student User Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "NEEDS_RESUBMISSION",
              "REJECTED",
              "EXPIRED"
            ],
            "title": "Status"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          },
          "assigned_guidance_staff_user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Guidance Staff User Id"
          },
          "decision_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision At"
          },
          "reviewed_by_user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By User Id"
          },
          "reason_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Code"
          },
          "reviewer_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Note"
          },
          "valid_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid Until"
          },
          "email_queued": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email Queued"
          }
        },
        "type": "object",
        "required": [
          "verification_id",
          "student_user_id",
          "status",
          "submitted_at",
          "assigned_guidance_staff_user_id",
          "decision_at",
          "reviewed_by_user_id",
          "reason_code",
          "reviewer_note",
          "valid_until"
        ],
        "title": "VerificationResponse"
      }
    }
  }
}